Update README.md
#Monte Carlo Simulations in Fortran

This repository contains a collection of Fortran routines that use Monte Carlo methods to perform numerical estimations. It includes subroutines for estimating the value of π and evaluating definite integrals, leveraging both the Hit-or-Miss and Crude Monte Carlo methods. The code also demonstrates the Rejection method for generating points from a specific distribution function.